% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox1973doc.R
\docType{data}
\name{wilcox1973}
\alias{wilcox1973}
\title{Results of the 1968 US presidential election}
\format{a \code{data_frame} with 52 rows and 4 variables.}
\source{
Wilcox, Allen R. 'Indices of Qualitative Variation and Political Measurement.' \emph{The Western Political Quarterly} 26, no. 2 (1 June 1973): 325-43. doi:10.2307/446831.
}
\usage{
wilcox1973
}
\description{
A dataset containing the returns of the 1968 US presidential election, by states.
}
\keyword{datasets}
